DPL: introduce Analysis/DataModel
We should move here all the basic building blocks of our (ALICE)
AOD schema. I also extracted Gian Michele SecondaryVertex schema
from the task and put it here.
