+++ Analysis/DataModel/src/dummy.cxx
@@ -1,0 +1,1 @@
+#include "Analysis/SecondaryVertex.h"
