+++ Analysis/Tasks/CMakeLists.txt
@@ -20,7 +20,7 @@ o2_add_executable(correlations-collection
                   
 o2_add_executable(vertexing-hf
 	  	  SOURCES vertexerhf.cxx
-		  PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase 
+		  PUBLIC_LINK_LIBRARIES O2::Framework O2::AnalysisDataModel O2::DetectorsBase 
 	          COMPONENT_NAME Analysis)
 
 o2_add_executable(validation
